requirements.txt
#-e git+https://github.com/rst2pdf/rst2pdf.git@ef8183fc8003560274d48159c3741a8f04b40efb#egg=rst2pdf
Babel>=2.3,<3.0
Jinja2>=2.8,<3.0
MarkupSafe>=0.23,<1.0
Pillow>=2.9.0,<3.0
PyYAML>=3.11,<4.0
Pygments>=2.0.2,<3.0
Sphinx==1.4.4
alabaster==0.7.8
astroid==1.4.5
docutils>=0.12,<1.0
ipython
pdfrw>=0.2,<1.0
pylint>=1.6,<1.7
pytz>=2015.4,<2016.0
reportlab>=3.2.0,<4.0
rst2pdf>=0.93,<1.0
six>=1.9.0,<2.0
snowballstemmer>=1.2.0,<2.0
sphinx-rtd-theme>=0.1.9,<1.0
svglib>=0.6.3,<1.0
